Draw board and walls

+++ project/Model/Point2D.cpp
@@ -4,28 +4,28 @@
 
 #include "Point2D.h"
 
-Point2D::Point2D(int x, int y){
+Point2D::Point2D(float x, float y){
     this->x = x;
     this->y = y;
 }
 
-int Point2D::getX() {
+float Point2D::getX() {
     return x;
 }
 
-int Point2D::getY() {
+float Point2D::getY() {
     return y;
 }
 
-void Point2D::setX(int x) {
+void Point2D::setX(float x) {
     this->x = x;
 }
 
-void Point2D::setY(int y) {
+void Point2D::setY(float y) {
     this->y = y;
 }
 
-std::pair<int,int> Point2D::getPosition() {
+std::pair<float,float> Point2D::getPosition() {
     return std::make_pair(x,y);
 }
